determined that application/json-responding APIs in Web API quote outputted strings, whereas minimal API does not.
Also ascertained that minimal API will block a post sent to a put, which is good news.

Also validated that minimal API fails to respond with a 415 when a POST request is sent to an endpoint expecting a JSON request, as does Web API. Instead, minimal throws. This is documented in the internal notes @bradygaster is taking on the journey exploring the side-by-side scenarios.